integration of feature extraction

+++ collector/features.cpp
@@ -4,7 +4,7 @@
 void Recog::Features::extract(cv::Mat x, Recog::Wavelet w) {
     unsigned N = x.cols;
     unsigned S = scales_.size();
-    cv::Mat features = cv::Mat::zeros(S, N, CV_32F);
+    cv::Mat features = cv::Mat::zeros(S, N, CV_32S);
 
     // iterate over shifts (rows)
     for(unsigned n = 0; n < N; n++) {
@@ -13,7 +13,7 @@ void Recog::Features::extract(cv::Mat x, Recog::Wavelet w) {
             w.setScale(s);
             w.setShift(n);
             // convolution
-            features.at<float>(s,n) = w*x;
+            features.at<int>(s,n) = w*x;
         }
     }
     extracted_ = true;
@@ -26,7 +26,7 @@ double Recog::Wavelet::operator*(const cv::Mat& x) {
         unsigned sigPos = T-(t+1);
         // std::cerr << "x[T - t] = " << x.at<float>(sigPos) << "\n";
         // std::cerr << "Psi[t] = " << f_(t) << "\n";
-        sum += x.at<float>(sigPos) * f_(t + shift_);
+        sum += x.at<int>(sigPos) * f_(t + shift_);
     }
     return sum;
 }
